+++ src/utils.js
@@ -129,6 +129,15 @@ export function isDeclaredLostItem(item) {
   return get(item, 'status.name') === itemStatuses.DECLARED_LOST;
 }
 
+export function isWithdrawnItem(item) {
+  return get(item, 'status.name') === itemStatuses.WITHDRAWN;
+}
+
+export function isClaimedReturned(item) {
+  return get(item, 'status.name') === itemStatuses.CLAIMED_RETURNED;
+}
+
+
 export function isDelivery(request) {
   return get(request, 'fulfilmentPreference') === fulfilmentTypeMap.DELIVERY;
 }